Added an additional account to keep reward scores for a reporter in a network

+++ programs/hapi-core/src/state/community.rs
@@ -9,7 +9,7 @@ pub struct Community {
     pub cases: u64,
 
     /// Number of confirmations needed for address to be considered confirmed
-    pub confirmation_threshold: u32,
+    pub confirmation_threshold: u8,
 
     /// Number of epochs reporter must wait to retrieve their stake
     pub stake_unlock_epochs: u64,
